kube-manager/requirements.txt
django
djangorestframework
django-rest-swagger
ConfigParser
configparser
ruamel_yaml
coreapi
django-sslserver
pytz
pyping
kubernetes